Rename subset and intersection due to name clash
subset / 2 and intersection / 2 are already provided by SWI Prolog, making it
harder to effectively test the answers to the questions. Thus, the
documentation and the unit tests have been renamed to use the predicates
my_subset / 2 and my_intersection / 2 (essentially pseudo-namespacing to avoid
naming clashes). Thanks Prolog! :(

Additionally, a minor typo in the documentation for my_intersection has been
fixed.